Corrigindo erro em tabelas

+++ database/migrations/2026_05_23_005401_create_agendamentos_table.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('agendamentos', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->date('data');
-            $table->time('horario');
-            $table->string('servico');
-            $table->timestamps();
+        if (! Schema::hasTable('agendamentos')) {
+            Schema::create('agendamentos', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->date('data');
+                $table->time('horario');
+                $table->string('servico');
+                $table->timestamps();
 
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-        });
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+            });
+        }
     }
 
     /**
